1.001
Are you able to change the site so all gallery images open up in a
Light box?
http://jwms.com.au/jetwave-orcadia/

Also, Can the thumbnail images which show on the Vessels and Home pages
be a different image to the Feature image, then can you add the images
which are in the design file

Also, are you able to remove the side bar, on all post pages.